+++ js/arrange.js
@@ -13,15 +13,15 @@
 }(function($) {
 /* end-amd-strip-block */
 
-  $.fn.sort = function(options) {
+  $.fn.arrange = function(options) {
     'use strict';
 
     // Settings and Options
-    var pluginName = 'sort',
+    var pluginName = 'arrange',
         defaults = {
           itemsSelector: null,
           connectWith: false,
-          placeholderCssClass: 'sort-placeholder'
+          placeholderCssClass: 'arrange-placeholder'
         },
         settings = $.extend({}, defaults, options);
 
@@ -41,37 +41,37 @@
       },
 
       // Handle Events
-      // example from: https://github.com/farhadi/html5sortable/blob/master/jquery.sortable.js
+      // example from: https://github.com/farhadi/html5arrangeable/blob/master/jquery.arrangeable.js
       handleEvents: function() {
         var self = this,
           index, isHandle,
           status = {},
-          items = self.element.children().not('[data-sort-exclude="true"]'),
+          items = self.element.children().not('[data-arrange-exclude="true"]'),
           placeholder = $('<' + (/^(ul|ol)$/i.test(self.element[0].tagName) ? 'li' : 'div') +'>');
 
 
         if (settings.itemsSelector) {
-          items = $(settings.itemsSelector, self.element).not('[data-sort-exclude="true"]');
+          items = $(settings.itemsSelector, self.element).not('[data-arrange-exclude="true"]');
           placeholder = $('<'+ items.first()[0].tagName +' />');
         }
 
-        self.dragStart = 'dragstart.sort touchstart.sort gesturestart.sort';
-        self.dragEnd = 'dragend.sort touchend.sort touchcancel.sort gestureend.sort';
-        self.dragWhileDragging = 'dragover.sort dragenter.sort drop.sort touchmove.sort gesturechange.sort';
+        self.dragStart = 'dragstart.arrange touchstart.arrange gesturestart.arrange';
+        self.dragEnd = 'dragend.arrange touchend.arrange touchcancel.arrange gestureend.arrange';
+        self.dragWhileDragging = 'dragover.arrange dragenter.arrange drop.arrange touchmove.arrange gesturechange.arrange';
 
-        self.handle = self.element.attr('data-sort-handle');
-        self.connectWith = self.element.attr('data-sort-connectWith');
+        self.handle = self.element.attr('data-arrange-handle');
+        self.connectWith = self.element.attr('data-arrange-connectWith');
         self.placeholders = placeholder.addClass(settings.placeholderCssClass +' draggable');
 
         // Use Handle if available
         $(self.handle, items).addClass('draggable')
-          .on('mousedown.sort touchstart.sort', function() { isHandle = true; })
-          .on('mouseup.sort touchend.sort', function() { isHandle = false; });
+          .on('mousedown.arrange touchstart.arrange', function() { isHandle = true; })
+          .on('mouseup.arrange touchend.arrange', function() { isHandle = false; });
 
         // Add connect with
         if (self.connectWith) {
           items = items
-            .add($(self.connectWith).children().not('[data-sort-exclude="true"]'))
+            .add($(self.connectWith).children().not('[data-arrange-exclude="true"]'))
             .data('connectWith', self.connectWith);
         }
 
@@ -79,7 +79,7 @@
         items
         .attr('draggable', true).addClass(self.handle ? '' : 'draggable')
         .add([this, placeholder])
-        .not('a[href], img').on('selectstart.sort', function() {
+        .not('a[href], img').on('selectstart.arrange', function() {
           if(this.dragDrop) {
             this.dragDrop();//ie9
           }
@@ -89,17 +89,17 @@
         .each(function() {
           $(this)
           // Drag start --------------------------------------------------------------------------
-          .on(self.dragStart, function(e) {          
+          .on(self.dragStart, function(e) {
             if (self.handle && !isHandle) {
               return false;
             }
             isHandle = false;
             self.dragging = $(this);
 
-            index = self.dragging.addClass('sort-dragging').index();
+            index = self.dragging.addClass('arrange-dragging').index();
 
             $.extend(status, {start: self.dragging, startIndex: index});
-            self.element.triggerHandler('beforesort', status);
+            self.element.triggerHandler('beforearrange', status);
 
             var dt = e.originalEvent.dataTransfer;
             dt.effectAllowed = 'move';
@@ -112,12 +112,12 @@
               return;
             }
             self.placeholders.filter(':visible').after(self.dragging);
-            self.dragging.removeClass('sort-dragging').show();
+            self.dragging.removeClass('arrange-dragging').show();
             self.placeholders.detach();
 
             if (index !== self.dragging.index()) {
               $.extend(status, {end: self.dragging, endIndex: self.dragging.index()});
-              self.element.triggerHandler('sortupdate', status);
+              self.element.triggerHandler('arrangeupdate', status);
             }
             self.dragging = null;
           })
@@ -127,10 +127,10 @@
             var overItem = this,
               overIndex;
             e.preventDefault();
-            
+
             if(e.type==='drop') {
               e.stopPropagation();
-              self.dragging.trigger('dragend.sort');
+              self.dragging.trigger('dragend.arrange');
               return false;
             }
 
@@ -157,7 +157,7 @@
               }
 
               $.extend(status, {over: overItem, overIndex: overIndex});
-              self.element.triggerHandler('draggingsort', status);
+              self.element.triggerHandler('draggingarrange', status);
 
               // Fix: IE-11 on windows-10 svg was disappering
               var svg = $('svg', overItem);
@@ -194,8 +194,8 @@
         var items = (this.connectWith) ?
           this.element.children().add($(this.connectWith).children()) : this.element.children();
 
-        items.off('selectstart.sort '+ this.dragStart +' '+ this.dragEnd +' '+ this.dragWhileDragging);
-        $(this.handle, items).off('mousedown.sort mouseup.sort touchstart.sort touchend.sort');
+        items.off('selectstart.arrange '+ this.dragStart +' '+ this.dragEnd +' '+ this.dragWhileDragging);
+        $(this.handle, items).off('mousedown.arrange mouseup.arrange touchstart.arrange touchend.arrange');
         $.removeData(this.element[0], pluginName);
       }
     };